# apis/espresso/docs/swagger.yaml
definitions:
  router.ErrorResponse:
    properties:
      error:
        example: 'invalid id: not-a-uuid'
        type: string
    type: object
...
  router.Signal:
    properties:
      briefing:
        example: On 2026-06-02, U.S. lawmakers and the Trump administration debated
          AI sovereign-wealth and compute-tax proposals amid soaring inflation...
        type: string
      drivers:
        example:
        - High inflation and rising consumer costs driven by supply-chain bottlenecks
          and geopolitical tensions.
        items:
          type: string
        type: array
      events:
        example:
        - '2026-06-01: Senator Bernie Sanders introduced a 50% ownership tax on major
          AI firms'
        items:
          type: string
        type: array
      forecast:
        example: 'Short-term: Market volatility will persist, AI regulatory scrutiny
          will intensify, and consumer confidence remains low.'
        type: string
      id:
        example: e7d7571a-13f0-56f0-8563-50863b79c781
        format: uuid
        type: string
      impact_level:
        enum:
        - low
        - medium
        - high
        example: high
        type: string
      impacted_domains:
        example:
        - finance
        - technology
        - cybersecurity
        - labor
        - healthcare
        - energy
        - policy
        items:
          type: string
        type: array
      impacts:
        example:
        - 9.3% market sell-off across tech and financial sectors.
        - Decline in consumer confidence and increased credit-card delinquency.
        items:
          type: string
        type: array
      reported:
        example: "2026-06-02T14:02:00-04:00"
        type: string
      tags:
        example:
        - ai_sovereign_wealth_fund
        - ai_taxation
        - compute_tax
        - inflation
        - market_volatility
        items:
          type: string
        type: array
    type: object
info:
  contact:
    email: soumitsrah@cafecito.tech
    name: Project Cafecito
    url: http://cafecito.tech
  description: |-
    MCP-ready business intelligence over curated "sips" for agents, dashboards, and automated research workflows.
    A **sip** is one unit of intelligence. `event` records describe observed business or market developments. `signal` records synthesize higher-level implications from related events and actions. `action` records are lower-level source facts used by the ingestion pipeline.
    Agent workflow: (1) listTags to discover filter vocabulary; (2) searchEvents for time-ordered developments; (3) searchSignals for synthesized implications; (4) getRelatedSips to follow `same_as` duplicates or `derived_from` intelligence chains.
    Conventions: Auth is optional at the backend but API-key protected through the gateway. Pagination uses `limit` default 16 max 128 and `offset` default 0. Empty result sets return HTTP 204, not an error. All sip IDs are UUID strings such as `339366bc-464d-582f-8132-6875ccc814d2`.
    Response formats: use `response_type=json` for structured application data. Use `response_type=text` for MCP/LLM context; it returns the same underlying records as compact field-per-line plain text with lower token overhead.
  license:
    name: MIT
  title: Espresso API & MCP
  version: "0.1"
paths:
  /events:
    get:
      description: |-
        Returns event-kind sips sorted by `created` descending, newest first.
        **When to use**: retrieve concrete developments, incidents, company actions, policy changes, market moves, or other observed business events before moving to higher-level signals.
        **Search modes**: use `ids` for exact UUID lookup, `tags` for inclusive-AND tag filtering, `q` + `acc` for semantic search, and `from` to set the oldest creation date. These filters can be combined.
        **Default time window**: when `from` is omitted, the service uses its default recent window, currently about the last 7 days.
        **Response shape**: JSON responses are flattened digest objects with `id` and `reported` added by the router. Stable fields include `briefing`, `event_type`, `actions`, `people`, `regions`, `cross_domain_impacts`, `future_outlook`, `impact_level`, and `tags`; additional pipeline-specific keys may appear.
        **Agent format**: use `response_type=text` for compact field-per-line records when feeding an LLM or MCP client. Use JSON when the caller needs structured parsing.
      operationId: searchEvents
      parameters:
      - collectionFormat: csv
        description: Exact event sip UUIDs to fetch (CSV). Use when following references
          or retrieving known records.
        in: query
        items:
          type: string
        name: ids
        type: array
      - collectionFormat: csv
        description: Tag filters (CSV). Multiple values are inclusive AND, so every
          supplied tag must match.
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Natural-language semantic search query. Max 1024 characters;
          requires the embedder.
        in: query
        maxLength: 1024
        name: q
        type: string
      - default: 0.75
        description: Match strictness for q. 0.0=broad, 1.0=strict. Default 0.75.
        in: query
        maximum: 1
        minimum: 0
        name: acc
        type: number
      - description: Only include events created on or after this date (YYYY-MM-DD).
          Defaults to the recent window when omitted.
        format: date
        in: query
        name: from
        type: string
      - default: json
        description: Output format. json returns flattened digest objects; text returns
          compact plain-text records for LLM/MCP context.
        enum:
        - json
        - text
        in: query
        name: response_type
        type: string
      - default: 16
        description: Page size. Default 16, max 128.
        in: query
        maximum: 128
        minimum: 1
        name: limit
        type: integer
      - description: Number of events to skip. Default 0.
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: Event digests when response_type=json; plain-text event blocks
            when response_type=text
          schema:
            items:
              $ref: '#/definitions/router.Event'
            type: array
        "204":
          description: No matching events (empty result, not an error)
        "400":
          description: Invalid query parameters or malformed UUID in ids
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "429":
          description: Concurrency limit exceeded; retry shortly
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Database or embedder unavailable; retry
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Search event intelligence
      tags:
      - Events
...
  /related/{relationship}:
    get:
      description: |-
        Returns sips linked to one or more source UUIDs through the requested relationship.
        **When to use**: after searchEvents or searchSignals, call this endpoint to expand context around a known sip, deduplicate equivalent records, or trace derived intelligence.
        **Relationships**: `same_as` returns equivalent or duplicate records. `derived_from` returns downstream records generated from, or based on, the supplied source sip IDs.
        **Input**: `ids` is required and must contain one or more RFC 4122 UUID strings. The `relationship` path value must be exactly `same_as` or `derived_from`.
        **Response shape**: each result is a flattened event or signal digest with `id` and `reported` added by the router. Additional pipeline-specific keys may appear.
        **Agent format**: use `response_type=text` for compact field-per-line related records; use JSON when the caller needs structured parsing.
      operationId: getRelatedSips
      parameters:
      - description: Relationship to traverse. same_as finds equivalent records; derived_from
          follows generated intelligence.
        enum:
        - same_as
        - derived_from
        in: path
        name: relationship
        required: true
        type: string
      - collectionFormat: csv
        description: 'Source sip UUIDs (CSV). Example: b07049b5-54c0-50b0-a620-d3aea3f8a173'
        in: query
        items:
          type: string
        name: ids
        required: true
        type: array
      - default: json
        description: Output format. json returns flattened digest objects; text returns
          compact plain-text records for LLM/MCP context.
        enum:
        - json
        - text
        in: query
        name: response_type
        type: string
      - default: 16
        description: Page size. Default 16, max 128.
        in: query
        maximum: 128
        minimum: 1
        name: limit
        type: integer
      - description: Number of related records to skip. Default 0.
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: Related event/signal digests when response_type=json; plain-text
            related record blocks when response_type=text
          schema:
            items:
              $ref: '#/definitions/router.Event'
            type: array
        "204":
          description: No related sips found (empty result, not an error)
        "400":
          description: Missing ids, invalid relationship, invalid response_type, or
            malformed UUID
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "429":
          description: Concurrency limit exceeded; retry shortly
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Database unavailable; retry
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Follow related intelligence records
      tags:
      - Related
  /signals:
    get:
      description: |-
        Returns signal-kind sips sorted by `created` descending, newest first.
        **When to use**: retrieve synthesized business implications, forecasts, drivers, and cross-event patterns after or instead of searching raw events.
        **Search modes**: use `ids` for exact UUID lookup, `tags` for inclusive-AND tag filtering, `q` + `acc` for semantic search, and `from` to set the oldest creation date. These filters can be combined.
        **Default time window**: when `from` is omitted, the service uses its default recent window, currently about the last 7 days.
        **Response shape**: JSON responses are flattened digest objects with `id` and `reported` added by the router. Stable fields include `briefing`, `events`, `drivers`, `impacts`, `impacted_domains`, `forecast`, `impact_level`, and `tags`; additional pipeline-specific keys may appear.
        **Agent format**: use `response_type=text` for compact field-per-line records when feeding an LLM or MCP client. Use JSON when the caller needs structured parsing.
      operationId: searchSignals
      parameters:
      - collectionFormat: csv
        description: Exact signal sip UUIDs to fetch (CSV). Use when following references
          or retrieving known records.
        in: query
        items:
          type: string
        name: ids
        type: array
      - collectionFormat: csv
        description: Tag filters (CSV). Multiple values are inclusive AND, so every
          supplied tag must match.
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Natural-language semantic search query. Max 1024 characters;
          requires the embedder.
        in: query
        maxLength: 1024
        name: q
        type: string
      - default: 0.75
        description: Match strictness for q. 0.0=broad, 1.0=strict. Default 0.75.
        in: query
        maximum: 1
        minimum: 0
        name: acc
        type: number
      - description: Only include signals created on or after this date (YYYY-MM-DD).
          Defaults to the recent window when omitted.
        format: date
        in: query
        name: from
        type: string
      - default: json
        description: Output format. json returns flattened digest objects; text returns
          compact plain-text records for LLM/MCP context.
        enum:
        - json
        - text
        in: query
        name: response_type
        type: string
      - default: 16
        description: Page size. Default 16, max 128.
        in: query
        maximum: 128
        minimum: 1
        name: limit
        type: integer
      - description: Number of signals to skip. Default 0.
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      - text/plain
      responses:
        "200":
          description: Signal digests when response_type=json; plain-text signal blocks
            when response_type=text
          schema:
            items:
              $ref: '#/definitions/router.Signal'
            type: array
        "204":
          description: No matching signals (empty result, not an error)
        "400":
          description: Invalid query parameters or malformed UUID in ids
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "429":
          description: Concurrency limit exceeded; retry shortly
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Database or embedder unavailable; retry
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Search synthesized signals
      tags:
      - Signals
...
schemes:
- https
swagger: "2.0"